requirements.txt
streamlit==1.45.0
requests==2.31.0
python-dotenv==1.0.1
pandas==2.2.1
numpy==1.26.4 